collabsphere/backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Settings & Environment
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Real-time & WebSocket
python-socketio==5.10.0
redis==5.0.1

# AWS & AI
boto3==1.34.14
httpx==0.26.0

# File Storage
cloudinary==1.38.0

# Email
aiosmtplib==3.0.1

# Async utilities
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
